+++ game/end_to_end_tests/base_game_test.py
@@ -40,8 +40,6 @@ import time
 from unittest import skipUnless
 
 from django.core.urlresolvers import reverse
-from django.contrib.staticfiles.testing import LiveServerTestCase
-from django.conf import settings
 from django_selenium_clean import selenium
 
 from . import custom_handler
@@ -49,6 +47,7 @@ from portal.models import UserProfile
 from game.models import Workspace
 from .editor_page import EditorPage
 from .game_page import GamePage
+from .selenium_test_case import SeleniumTestCase
 from portal.tests.pageObjects.portal.home_page import HomePage
 from portal.tests.utils.organisation import create_organisation_directly
 from portal.tests.utils.teacher import signup_teacher_directly
@@ -57,28 +56,6 @@ from portal.tests.utils.student import create_school_student_directly
 
 
 custom_handler.monkey_patch()
-
-class SeleniumTestCase(LiveServerTestCase):
-
-    @classmethod
-    def setUpClass(cls):
-
-        super(SeleniumTestCase, cls).setUpClass()
-
-        # Normally we would just do something like
-        #     selenium.live_server_url = self.live_server_url
-        # However, there is no "self" at this time, so we
-        # essentially duplicate the code from the definition of
-        # the LiveServerTestCase.live_server_url property.
-        selenium.live_server_url = 'http://%s:%s' % (
-            cls.server_thread.host, cls.server_thread.port)
-
-    def __call__(self, result=None):
-        if not selenium:
-            return super(SeleniumTestCase, self).__call__(result)
-        for width in getattr(settings, 'SELENIUM_WIDTHS', [1024]):
-            selenium.set_window_size(width, 1024)
-            super(SeleniumTestCase, self).__call__(result)
 
 
 @skipUnless(selenium, "Selenium is unconfigured")
