+++ setup.py
@@ -35,6 +35,7 @@ setup(
         "six==1.11.0",
         "docutils==0.12",
         "pyhamcrest==1.8.3",
+        "libsass",
     ],
     tests_require=["django-selenium-clean==0.2.1", "selenium==3.7.0"],
     test_suite="test_utils.test_suite.DjangoAutoTestSuite",
